refactor: now use single-thread-for-loop for collecting estates
