+++ batch/src/main/java/com/zipsoon/batch/estate/job/config/EstateStepConfig.java
@@ -1,15 +1,10 @@
-package com.zipsoon.batch.job.config;
+package com.zipsoon.batch.estate.job.config;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.zipsoon.batch.dto.NaverResponseDto;
-import com.zipsoon.batch.job.listener.EstateStepListener;
-import com.zipsoon.batch.job.partitioner.DongCodePartitioner;
-import com.zipsoon.batch.job.processor.EstateItemProcessor;
-import com.zipsoon.batch.job.reader.EstateItemReader;
-import com.zipsoon.batch.job.reader.JsonEstateItemReader;
-import com.zipsoon.batch.job.writer.EstateItemWriter;
-import com.zipsoon.batch.service.DongCodeService;
-import com.zipsoon.batch.service.NaverClient;
+import com.zipsoon.batch.estate.domain.DongCode;
+import com.zipsoon.batch.estate.job.listener.EstateStepListener;
+import com.zipsoon.batch.estate.job.processor.EstateItemProcessor;
+import com.zipsoon.batch.estate.job.writer.EstateItemWriter;
+import com.zipsoon.batch.estate.service.DongCodeService;
 import com.zipsoon.common.domain.EstateSnapshot;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,13 +12,13 @@ import org.springframework.batch.core.Step;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.builder.StepBuilder;
 import org.springframework.batch.item.ItemReader;
+import org.springframework.batch.item.support.ListItemReader;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.env.Environment;
 import org.springframework.transaction.PlatformTransactionManager;
 
-import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Configuration
 @RequiredArgsConstructor
@@ -31,19 +26,16 @@ import java.util.List;
 public class EstateStepConfig {
     private final JobRepository jobRepository;
     private final PlatformTransactionManager transactionManager;
-    private final NaverClient naverClient;
     private final DongCodeService dongCodeService;
     private final EstateItemProcessor estateItemProcessor;
     private final EstateItemWriter estateItemWriter;
-    private final ObjectMapper objectMapper;
-    private final Environment environment;
 
 
     @Bean
     public Step estateWorkerStep() {
         return new StepBuilder("estateWorkerStep", jobRepository)
-            .<NaverResponseDto, List<EstateSnapshot>>chunk(1, transactionManager)
-            .reader(selectItemReader())
+            .<String, List<EstateSnapshot>>chunk(1, transactionManager)
+            .reader(dongCodeReader())
             .processor(estateItemProcessor)
             .writer(estateItemWriter)
             .listener(new EstateStepListener())
@@ -51,22 +43,10 @@ public class EstateStepConfig {
     }
 
     @Bean
-    public DongCodePartitioner dongCodePartitioner() {
-        return new DongCodePartitioner(dongCodeService);
-    }
-
-    private ItemReader<NaverResponseDto> selectItemReader() {
-        if (isLocalProfile()) {
-            log.info("Using JsonEstateItemReader (mock data) for local profile");
-            return new JsonEstateItemReader(objectMapper);
-        } else {
-            log.info("Using EstateItemReader (real API call)");
-            return new EstateItemReader(naverClient, dongCodeService);
-        }
-    }
-
-    private boolean isLocalProfile() {
-        return Arrays.asList(environment.getActiveProfiles()).contains("local");
+    public ItemReader<String> dongCodeReader() {
+        return new ListItemReader<>(dongCodeService.getAllDongCodes().stream()
+            .map(DongCode::code)
+            .collect(Collectors.toList()));
     }
 
 }
